partially support function pointers in python
this is one-way only so far -- a function pointer (or the function
itself in python terms) can be passed around, but not invoked. (this is
useful when passing function pointers to the surrounding system by means
of user calls, for example).

in order to use function pointers, types should be implemented.